Admin refactor

+++ code/GeoQR/app/src/main/java/com/example/geoqr/Admin.java
@@ -41,7 +41,12 @@ public class Admin {
         // Adapters
         playerAdapter = new AdminPlayerAdapter(ctx, new ArrayList<>());
         qrAdapter = new AdminQRAdapter(ctx, new ArrayList<>());
+    }
 
+    /**
+     * It populates the Array Adapters by fetching data from the db
+     */
+    public void fetch() {
         // Get List of QR codes
         db = FirebaseFirestore.getInstance();
         db.collection("QR codes").get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
@@ -79,11 +84,14 @@ public class Admin {
      */
     public void deleteQRCodes() {
         for (AdminQRTuple qrTuple: qrSelection) {
-            db.collection("QR codes").document(qrTuple.getContents()).delete();
-            db.collection("User")
-                    .document(qrTuple.getPlayer())
-                    .collection("QR codes").document(qrTuple.getContents())
-                    .delete();
+            if (db != null) {
+                db.collection("QR codes").document(qrTuple.getContents()).delete();
+                db.collection("User")
+                        .document(qrTuple.getPlayer())
+                        .collection("QR codes").document(qrTuple.getContents())
+                        .delete();
+            }
+
             qrAdapter.remove(qrTuple);
         }
 
@@ -96,7 +104,10 @@ public class Admin {
      */
     public void deletePlayers() {
         for (AdminPlayerTuple playerTuple: playerSelection) {
-            db.collection("Users").document(playerTuple.getName()).delete();
+            if (db != null) {
+                db.collection("Users").document(playerTuple.getName()).delete();
+            }
+
             playerAdapter.remove(playerTuple);
         }
